requirements.txt
streamlit
pymongo
python-dotenv
langchain
langchain_groq
langchain_core